refactor: reorder UI logic to a proper dedicated folder
