+++ src/app/(ui)/SyncForm.tsx
@@ -4,7 +4,7 @@ import { PrimaryBtn } from '@/components/Buttons'
 import { Heading, SubHeading } from '@/components/Typography'
 import { FormBox, Label, TextInput, Selecter } from '@/components/FormElements'
 import { Box, Typography } from '@mui/material'
-import { SyncOption, syncConfigurationOptions, syncOptions } from '@/app/helpers'
+import { SyncOption, syncConfigurationOptions, syncOptions } from '@ui/helpers'
 import { DefaultSetting, SelecterOption } from '@/types/settings'
 import { Setting } from '@prisma/client'
 import { useFormState } from 'react-dom'
