+++ src/app/(ui)/ThemeRegistry.tsx
