+++ src/app/api/webhooks/copilot/route.ts
@@ -1,4 +1,4 @@
-import { withErrorHandler } from '@/app/api/core/utils/withErrorHandler'
+import { withErrorHandler } from '@ui/api/core/utils/withErrorHandler'
 import { handleCopilotWebhookEvent } from './copilot.controller'
 
 export const POST = withErrorHandler(handleCopilotWebhookEvent)
